+++ src/common/options.js
@@ -153,6 +153,22 @@ const options = {
         },
         events:{}
     },
+    selectedFacets:{
+        enabled:true,
+        selectedFacetClass:"UNX-selected-facet-btn",
+        el:null,
+        template:selectedFacetUI,
+        itemTemplate:selectedFacetItemTemplateUI,
+        clearAllText:"Clear All",
+        tagName:"DIV",
+        htmlAttributes:{
+            class:"UNX-selected-facet-lb"
+        },
+        events:{},
+        facetAction:"click",
+        clearFacetsSelectorClass:'UNX-clear-facet',
+        removeFacetsSelectorClass:'UNX-remove-facet',
+    },
 
 
     facet: {
@@ -165,17 +181,7 @@ const options = {
 
         selectedFacetClass:"UNX-selected-facet-btn",
         selectedFacetsEl:null,
-        selectedFacetTemplate:selectedFacetUI,
-        selectedFacetItemTemplate:selectedFacetItemTemplateUI,
         clearAllText:"Clear All",
-        selectedFacetConfig:{
-            tagName:"DIV",
-            htmlAttributes:{
-                class:"UNX-selected-facet-lb"
-            },
-            events:{}
-        },
-
 
         rangeTemplate:renderRangeFacets,
         rangeWidgetConfig: {