+++ CHANGELOG.md
@@ -18,13 +18,13 @@
 - Added method to get all the selected facets - instance.getAllSelectedFacets()
 - Added a method to get the total count of selected facets - instance.getSelectedFacetsCount()
 - Added a method to get the total count of selected facet values - instance.getSelectedFacetValueCount()
-- Added tagName, events and htmlAttributes to each config so that wrapper element created by sdk can be easily customised by customer
-- Added browseQueryParam to set the default category param for the category/ browse calls.
-
+- Added selectedFacets config to manage selected facets, its seperated from facet config
+- Added extra parameters to selected facet template and item template methods
 
 ### Changed
 - fixed the special character issue in multilevel facets.
 - fixed the issue while getting selected multilevel facet method
+- deprecated the selected facets config from facet config.
 
 ## [2.0.0] - 2020-09-28